+++ hspec-core/test/Test/Hspec/Core/OptionsSpec.hs
@@ -18,58 +18,58 @@ spec = do
     let parseOptions = Options.parseOptions defaultConfig "my-spec"
 
     it "rejects unexpected arguments" $ do
-      fromLeft (parseOptions [] ["foo"]) `shouldBe` (ExitFailure 1, "my-spec: unexpected argument `foo'\nTry `my-spec --help' for more information.\n")
+      fromLeft (parseOptions [] Nothing ["foo"]) `shouldBe` (ExitFailure 1, "my-spec: unexpected argument `foo'\nTry `my-spec --help' for more information.\n")
 
     it "rejects unrecognized options" $ do
-      fromLeft (parseOptions [] ["--foo"]) `shouldBe` (ExitFailure 1, "my-spec: unrecognized option `--foo'\nTry `my-spec --help' for more information.\n")
+      fromLeft (parseOptions [] Nothing ["--foo"]) `shouldBe` (ExitFailure 1, "my-spec: unrecognized option `--foo'\nTry `my-spec --help' for more information.\n")
 
     it "sets configColorMode to ColorAuto" $ do
-      configColorMode <$> parseOptions [] [] `shouldBe` Right ColorAuto
+      configColorMode <$> parseOptions [] Nothing [] `shouldBe` Right ColorAuto
 
     context "with --no-color" $ do
       it "sets configColorMode to ColorNever" $ do
-        configColorMode <$> parseOptions [] ["--no-color"] `shouldBe` Right ColorNever
+        configColorMode <$> parseOptions [] Nothing ["--no-color"] `shouldBe` Right ColorNever
 
     context "with --color" $ do
       it "sets configColorMode to ColorAlways" $ do
-        configColorMode <$> parseOptions [] ["--color"] `shouldBe` Right ColorAlways
+        configColorMode <$> parseOptions [] Nothing ["--color"] `shouldBe` Right ColorAlways
 
     context "with --out" $ do
       it "sets configOutputFile" $ do
-        either (const Nothing) Just . configOutputFile <$> parseOptions [] ["--out", "foo"] `shouldBe` Right (Just "foo")
+        either (const Nothing) Just . configOutputFile <$> parseOptions [] Nothing ["--out", "foo"] `shouldBe` Right (Just "foo")
 
     context "with --qc-max-success" $ do
       context "when given an invalid argument" $ do
         it "returns an error message" $ do
-          fromLeft (parseOptions [] ["--qc-max-success", "foo"]) `shouldBe` (ExitFailure 1, "my-spec: invalid argument `foo' for `--qc-max-success'\nTry `my-spec --help' for more information.\n")
+          fromLeft (parseOptions [] Nothing ["--qc-max-success", "foo"]) `shouldBe` (ExitFailure 1, "my-spec: invalid argument `foo' for `--qc-max-success'\nTry `my-spec --help' for more information.\n")
 
     context "with --depth" $ do
       it "sets depth parameter for SmallCheck" $ do
-        configSmallCheckDepth <$> parseOptions [] ["--depth", "23"] `shouldBe` Right 23
+        configSmallCheckDepth <$> parseOptions [] Nothing ["--depth", "23"] `shouldBe` Right 23
 
     context "with --jobs" $ do
       it "sets number of concurrent jobs" $ do
-        configConcurrentJobs <$> parseOptions [] ["--jobs=23"] `shouldBe` Right (Just 23)
+        configConcurrentJobs <$> parseOptions [] Nothing ["--jobs=23"] `shouldBe` Right (Just 23)
 
       it "rejects values < 1" $ do
         let msg = "my-spec: invalid argument `0' for `--jobs'\nTry `my-spec --help' for more information.\n"
-        void (parseOptions [] ["--jobs=0"]) `shouldBe` Left (ExitFailure 1, msg)
+        void (parseOptions [] Nothing ["--jobs=0"]) `shouldBe` Left (ExitFailure 1, msg)
 
     context "when given a config file" $ do
       it "uses options from config file" $ do
-        configColorMode <$> parseOptions [("~/.hspec", ["--no-color"])] [] `shouldBe` Right ColorNever
+        configColorMode <$> parseOptions [("~/.hspec", ["--no-color"])] Nothing [] `shouldBe` Right ColorNever
 
       it "gives command-line options precedence" $ do
-        configColorMode <$> parseOptions [("~/.hspec", ["--no-color"])] ["--color"] `shouldBe` Right ColorAlways
+        configColorMode <$> parseOptions [("~/.hspec", ["--no-color"])] Nothing ["--color"] `shouldBe` Right ColorAlways
 
       it "rejects --help" $ do
-        fromLeft (parseOptions [("~/.hspec", ["--help"])] []) `shouldBe` (ExitFailure 1, "my-spec: unrecognized option `--help' in config file ~/.hspec\n")
+        fromLeft (parseOptions [("~/.hspec", ["--help"])] Nothing []) `shouldBe` (ExitFailure 1, "my-spec: unrecognized option `--help' in config file ~/.hspec\n")
 
       it "rejects unrecognized options" $ do
-        fromLeft (parseOptions [("~/.hspec", ["--invalid"])] []) `shouldBe` (ExitFailure 1, "my-spec: unrecognized option `--invalid' in config file ~/.hspec\n")
+        fromLeft (parseOptions [("~/.hspec", ["--invalid"])] Nothing []) `shouldBe` (ExitFailure 1, "my-spec: unrecognized option `--invalid' in config file ~/.hspec\n")
 
       it "rejects ambiguous options" $ do
-        fromLeft (parseOptions [("~/.hspec", ["--qc-max-s"])] []) `shouldBe` (ExitFailure 1,
+        fromLeft (parseOptions [("~/.hspec", ["--qc-max-s"])] Nothing []) `shouldBe` (ExitFailure 1,
           unlines [
             "my-spec: option `--qc-max-s' is ambiguous; could be one of:"
           , "  -a N  --qc-max-success=N  maximum number of successful tests"
@@ -81,7 +81,17 @@ spec = do
 
     context "when given multiple config files" $ do
       it "gives later config files precedence" $ do
-        configColorMode <$> parseOptions [("~/.hspec", ["--no-color"]), (".hspec", ["--color"])] [] `shouldBe` Right ColorAlways
+        configColorMode <$> parseOptions [("~/.hspec", ["--no-color"]), (".hspec", ["--color"])] Nothing [] `shouldBe` Right ColorAlways
+
+    context "when given an environment variable" $ do
+      it "uses options from environment variable" $ do
+        configColorMode <$> parseOptions [] (Just ["--no-color"]) [] `shouldBe` Right ColorNever
+
+      it "gives command-line options precedence" $ do
+        configColorMode <$> parseOptions [] (Just ["--no-color"]) ["--color"] `shouldBe` Right ColorAlways
+
+      it "rejects unrecognized options" $ do
+        fromLeft (parseOptions [] (Just ["--invalid"]) []) `shouldBe` (ExitFailure 1, "my-spec: unrecognized option `--invalid' from environment variable HSPEC_OPTIONS\n")
 
   describe "ignoreConfigFile" $ around_ (withEnvironment []) $ do
     context "by default" $ do